requirements.txt
numpy 
pandas
pytorch 
torchvision 
urllib3 
ray
imgaug 
tensorboard 